// src/app/vo/interfaces/detalles.interface.ts
export interface FichaData {
  aIdioma?: string;
  aIdStock: string;
}

export interface FichaDto {

  idVehi: number;
  idStoc: number;

  marca: string;
  modelo: string;
  tipoVeh: string;

  chasis: string;
  motor: string;

  potMot: number;
  kilom: number;

  euronorma: string;

  fecMat: string;

  adblue: number;
  adr: string;

  garvo: string;
  precio: number;

  cabina: string;
  asiext: string;
  color: string;
  spoiler: string;

  tipoCaja: string;
  modeloCaja: string;

  numVel: number;

  marcaEje: string;
  confEje: string;

  disEje: number;

  ejeEle: string;
  numEje: number;
  ejtDir: string;

  mma: number;
  mmaEjd: number;
  tara: number;

  susDel: string;
  susTra: string;

  frenoAux: string;
  frenoPri: string;

  numDep: number;
  depCom: number;
  depCoa: number;

  tipoCarro: string;
  tipoPiso: string;
  portaPalet: string;

  fecAtp: string;
  volume: number;

  marcaCarro: string;

  intLar: number;
  intAnc: number;
  intAlt: number;

  extLar: number;
  extAnc: number;
  extAlt: number;

  tecElec: string;
  tecCor: string;

  friEqu: string;
  horFri: number;

  marcaFrio: string;
  friMod: string;

  conEle: string;
  terMog: string;

  tacografo: string;
  ebs: string;
  conVel: string;

  airAco: string;
  calefa: string;
  cliTec: string;

  radio: string;
  emisora: string;
  nevera: string;
  xenon: string;

  literas: number;

  hidraulica: string;
  ganchoRem: string;
  remolcador: string;

  grua: string;
  manGrua: string;
  modGru: string;

  cabRes: string;
  altCha: number;

  tipoChasis: string;

  plataforma: string;
  marcaPlataforma: string;
  plaMod: string;
  plaEle: string;

  lona: string;
  techo: string;
  puerta: string;

  diBeje: string;
  codCol: string;

  vendido: string;
  ciudad: string;
  provincia: string;

  urlVid: string;

  vellaAle: string;

  veMmNeu1: number;
  veMmNeu3: number;

  vpGarvo: string;

  vhModelo: string;

  llantas: string;
  numTas: number;

  vetdf: string;
  navega: string;

  preTv: string;

  fecItv: string;

  matric: string;

  recepci: string;

  concep: string;

  numLla: number;

  ejeGrua: string;
  vhTipVeh: string;

  plaTra: string;
  plaRet: string;

  marcaGrua: string;

  numBar: number;
  cajPla: number;
  cajMet: number;

  bidaGru: number;

  duOtra: string;
  cfgeje: string;

  estadoStock: string;

  img1?: string;
  img2?: string;
  img3?: string;
  img4?: string;
  img5?: string;
  img6?: string;
  img7?: string;
  img8?: string;
  img9?: string;
  img10?: string;

  imgCarroceria?: string;
  imgGrua?: string;
  imgPlataforma?: string;
}